core/requirements.txt
asgiref==3.3.4
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.0
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
importlib-metadata==4.0.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
packaging==20.9
pep517==0.8.2
Pillow==8.1.2
progress==1.5
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
toml==0.10.2
typing==3.7.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
zipp==3.4.1
